Added death frame when player dies, player will not be able to move (or go to the next level) for 60 frames. Player will then be teleported to starting location.
Added "start screen", which displays an image when the game is first started or when the game is reloaded.

The "start screen" will disappear when you click on the game and will not be displayed until the game is reloaded.

For now the "start screen" is a big "START", I'll change it to the image of "CLICK HERE" when we have it.

Changed game states to only have "start" and "playing".

Added the "start screen" image, Start.png.