Added player carrying fat king frames. Currently not activated by colliding with the fat king.
Added separate left and right death frames.

Added function to setup for fat king images.

Added function to initiate carrying fat king, changing the player image, player dimensions, player movement and jump speed.

Added function to clear boolean properties of the player.

Added function to check if the game is started for the first time and display a start image if so.

Changed names of some assets and folders.

Added alternate versions of the left and right death frames of player.

Fixed the fat king in the cage image to not have chains.